初回アラートが２度表示されないようにした
